Basic GICv3 functionality. Alongside a lot of smaller changes/fixes including making the device tree code work on big endian systems (by writing a version of the libc endian.h file) and changing some Kconfig stuff to make it cleaner for non-x64 configuration

+++ include/nautilus/thread.h
@@ -53,7 +53,7 @@ typedef uint64_t nk_stack_size_t;
     
 #include <nautilus/scheduler.h>
 
-#define CPU_ANY       -1
+#define CPU_ANY       (int)-1
 
 /* common thread stack sizes */
 #define TSTACK_DEFAULT 0  // will be 4K
